Create README.md
A simple Todo list web app based on Javascript. 
It has three tabs - Add Item, ToDo Tasks, and Completed.
Add Item - A task can be added to the list of To-Do Tasks.
ToDo Tasks - Lists the current list of todo of the user, has the option to mark it as completed and move to the completed tab or delete it from the list altogether.

Completed - Lists the completed tasks of the user and an option to delete it from the list.